feat: Allow cards to block other cards
Consider three cards:

1. card-A with an ID of `card-a-id`
1. card-B with an ID of `card-b-id`
1. card-C with an ID of `card-c-id`

The user can set card-B as blocking card-A and card-C by setting the
`FC_BLOCKED_BY` property on cards A and B to `card-b-id`.

Alternatively, the user could set card-A and card-B as blocking card-C
by setting the `FC_BLOCKED_BY` property on card-C to `card-a-id,card-b-id`.