upodates org-fc-show-latex to continue despite error
Motivation: I was having a hard time getting latex set up properly, but I wanted to work through my cards regardless. Rendering the latex was not a prerequisite to working through the cards, but I was unable to complete the cards because of the latex error. This prints the error as a message but does not exit the review session.