feat: Add default algo variable
**What?**

Allow users to set a default algorithm with `org-fc-default-algo`. We
expect the variable to be one of the symbols in `org-fc-algos` (for
example: `'sm2`).

**Why?**

My use case, and I suspect the primary use case, is to have a single
algorithm among all cards. It'd be nice to skip the prompt in this
situation.